+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Drop `CSR_ALLOW_EXECUTABLE_POLICY_OVERRIDE` requirement for root patching
   - Lowers default SIP Disabled value to 0xA03
 - Remove legacy GMUX patch set for MacBookPro5,x machines due to instabilites
+- Fix non-Metal acceleration crashing on 12.0.1
+  - Yes Apple adding a notch broke our accel patches
+- Fix non-Metal Control Center crashing on 12.0 Beta 10+
 
 ## 0.3.0
 - Fix Nvidia Tesla Acceleration in Monterey Beta 7+